Update:  Changed method name (objectQuery -> traverseWithQuery)
