+++ lib/index.js
@@ -9,7 +9,7 @@ module.exports = {
 	walk: require('./walk'),
 	concat: require('./concat'),
 	createHash: require('./createHash'),
-	objectQuery: require('./objectQuery')
+	traverseWithQuery: require('./traverseWithQuery')
 };
 
 global.print = module.exports.print;
