+++ lib/traverseWithQuery/index.js
@@ -1,13 +1,14 @@
 'use strict';
 
-module.exports = function(target, verifying) {
-	return dfs(target, verifying);
+module.exports = function(struct, query) {
+	return dfs(struct, query);
 };
 
-var dfs = function(target, verifying) {
-	let ret = target,
-		query = verifying.split('.'),
+var dfs = function(struct, query) {
+	let ret = struct,
 		i;
+
+	query = Array.isArray(query) ? query : query.split('.');
 
 	for (i = 0; i < query.length - 1; i++) {
 		if (ret === undefined) {
